W4D3/Q1/requirements.txt
fastmcp==2.10.1 
discord.py==2.3.2
python-dotenv==1.1.0
pytest==7.4.3
pytest-asyncio==0.21.1 